Rearrange cpp functions into separate files
Makes it easier to develop + test! This is a squashed commit of the following:

commit 8b7c1e48136893a360782bda42a719419a2746a9
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:48:01 2023 +0100

    update exports files

commit 8b5ceb2d98ddc00ce7a4c47fa19b7eb0cd450a80
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:47:47 2023 +0100

    rename samplers.cpp to nuts.cpp

commit 0f18e1c1be3b3ace481f763828afc057fb840926
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:47:22 2023 +0100

    clean up samplers.cpp

commit 347768dd35d1ecc88aef5553d1513b6766a06145
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:46:51 2023 +0100

    remove duplicate function

commit 2ccb9f4792321d5230ac0e1050a85fa4541942e6
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:46:22 2023 +0100

    move samplers to their own files

commit b01ab6c3d2836091487e0a3575968db4b3f188c1
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:45:24 2023 +0100

    move hmc-style utils elsewhere

commit ea863dd4f35a9e38c787daea385e274e16c675cc
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:43:51 2023 +0100

    move functions needed for mh-style samplers elsewhere

commit 08f19f0b456c630bba1917852435d7fbaa5dbdb3
Author: Lucas Castillo <lucas.castillo.marti@gmail.com>
Date:   Tue Oct 24 17:42:57 2023 +0100

    move pdf-managing functions elsewhere
