+++ R/RcppExports.R
@@ -1,20 +1,20 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-sampler_mh_cpp <- function(start, sigma_prop, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha = 0) {
-    .Call(`_samplr_sampler_mh_cpp`, start, sigma_prop, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha)
+sampler_hmc_cpp <- function(start, distr_name, distr_params, epsilon, L, iterations, isMix, weights, custom_func, useCustom) {
+    .Call(`_samplr_sampler_hmc_cpp`, start, distr_name, distr_params, epsilon, L, iterations, isMix, weights, custom_func, useCustom)
 }
 
 sampler_mc3_cpp <- function(start, nChains, sigma_prop, delta_T, swap_all, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha = 0) {
     .Call(`_samplr_sampler_mc3_cpp`, start, nChains, sigma_prop, delta_T, swap_all, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha)
 }
 
-sampler_hmc_cpp <- function(start, distr_name, distr_params, epsilon, L, iterations, isMix, weights, custom_func, useCustom) {
-    .Call(`_samplr_sampler_hmc_cpp`, start, distr_name, distr_params, epsilon, L, iterations, isMix, weights, custom_func, useCustom)
-}
-
 sampler_mc_rec_cpp <- function(start, nChains, delta_T, swap_all, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, epsilon, L, alpha) {
     .Call(`_samplr_sampler_mc_rec_cpp`, start, nChains, delta_T, swap_all, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, epsilon, L, alpha)
+}
+
+sampler_mh_cpp <- function(start, sigma_prop, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha = 0) {
+    .Call(`_samplr_sampler_mh_cpp`, start, sigma_prop, iterations, distr_name, distr_params, discreteValues, isMix, weights, custom_func, useCustom, alpha)
 }
 
 sampler_nuts_cpp <- function(start, distr_name, distr_params, epsilon, iterations, delta_max, isMix, weights, custom_func, useCustom) {
